+++ src/models/inputedgeprometheusinput.ts
@@ -153,7 +153,7 @@ export type InputEdgePrometheusInput = {
    */
   dimensionList?: Array<string> | undefined;
   /**
-   * Enable to use each metric name as the event field key (e.g. go_threads: 9) instead of the default _metric/_value format.
+   * When enabled, each metric name is used as the event field key (example: go_threads: 9) instead of the default _metric/_value format.
    */
   fieldPerMetric?: boolean | undefined;
   /**