+++ RELEASES.md
@@ -1579,3 +1579,13 @@ Based on:
 - [typescript v0.8.0-alpha.4] .
 ### Releases
 - [NPM v0.8.0-alpha.4] https://www.npmjs.com/package/cribl-control-plane/v/0.8.0-alpha.4 - .
+
+## 2026-05-09 02:52:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.8.0-alpha.5] .
+### Releases
+- [NPM v0.8.0-alpha.5] https://www.npmjs.com/package/cribl-control-plane/v/0.8.0-alpha.5 - .